+++ Assets/Scripts/Characters/General/HealthUI.cs
@@ -19,7 +19,7 @@ namespace Characters.General
             fill.color = gradient.Evaluate(1f);
         }
 
-        public void SetHealth(int health)
+        public void SetHealth(float health)
         {
             slider.value = health;
             fill.color = gradient.Evaluate(slider.normalizedValue);
